LayoutTests/fast/mediastream/RTCPeerConnection-events-expected.txt

Tests that RTCPeerConnection event callbacks are async so that for example close can be called safely. The order of the messages is very important.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gotStream was called.
PASS gotStream done.
PASS onNegotiationNeeded was called.
PASS onNegotiationNeeded done.
PASS onStateChange was called.
PASS pc.signalingState is 'closed'
PASS successfullyParsed is true

TEST COMPLETE
